Use navigator.usb.requestDevice, not getDevices
getDevices only works for the devices that have already been paired via
the requestDevice method. Originally this was working for me because my
test device had already been paired on localhost via another app, so I
never ran into the problem.

Also note that doing requestDevice in response to getting nothing in
getDevice won't work, because the browser will raise a SecurityError:
a function that will result in asking permissions from a user must be
initiated in direct response to a user's gesture.

Thanks to this patch, I was also able to remove the device chooser
dialog, as we now rely on the browser's own.

Fixes: #1