fix(console): remove the scrollbar
Ideally we'd be able to show the scrollbar only if necessary, i.e. when
the content is larger than the viewport, but hterm doesn't support this.
We can't style it anyway becasue it's in an `iframe`. I suppose it's
better to simply forego the scrollbar as the user can scroll with the
mouse wheel.

Fixes #7